requirements.txt
# ─── Job Intelligence Agent — Dependencies ───

# Scraping
python-jobspy>=1.1.75

# LLM (OpenRouter - OpenAI-compatible API)
openai>=1.30.0

# HTTP & Config
requests>=2.31.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0       # PostgreSQL driver (for Neon prod)

# API & Dashboard
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Preference Learning
numpy>=1.24.0
scikit-learn>=1.3.0

# Scheduling (for Render deployment)
apscheduler>=3.10.0
